% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray_curtis.r
\name{bc}
\alias{bc}
\title{Compute bray-curtis dissimilarity based on species abundances}
\usage{
bc(comm)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns); values are abundances. Column and row labels are highly recommended}
}
\value{
Site by site dissimilarity matrix
}
\description{
Compute bray-curtis dissimilarity based on species abundances
}
