% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{branch_adjacency}
\alias{branch_adjacency}
\title{Branch adjacency}
\usage{
branch_adjacency(x, branch.lengths = !is.null(x$edge.length))
}
\arguments{
\item{x}{Object of class phylo}

\item{branch.lengths}{Logical; if true, the values of the adjacency matrix will be the
branch lengths, otherwise they will be 1 or 0}
}
\value{
A branch by node adjacency matrix
}
\description{
Produce a branch adjacency matrix (i.e., branch by node)
}
\keyword{internal}
