% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_mbm.r
\name{ci_parametric}
\alias{ci_parametric}
\title{Parametric confidence intervals for mbm objects}
\usage{
ci_parametric(x, level)
}
\arguments{
\item{level}{Confidence level required}

\item{object}{A fit from an mbm object; first column should be the prediction and second the standard deviation}
}
\value{
A matrix with upper and lower confidence limits in columns and observations in rows
}
\description{
Computed via normal approximation.
}
