% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dendro.r
\name{combine_traits}
\alias{combine_traits}
\title{Process functional trait data, combining multiple protocols and multiple ways of coding}
\usage{
combine_traits(x, traits = 1, categories = 2, values = 3)
}
\arguments{
\item{x}{"Tall" dataframe with species and trait information}

\item{traits}{column number or name in x of trait names}

\item{categories}{column number or name for trait categories (i.e., the 'value' of text-coded variables)}

\item{values}{column number or name for trait values (i.e., numeric values)}
}
\value{
dataframe similar to x, but with re-coded traits and with numeric values for text-coded traits
}
\description{
Process functional trait data, combining multiple protocols and multiple ways of coding
}
