% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_mbm.r
\name{confint.mbm}
\alias{confint.mbm}
\title{Confidence intervals for MBM models}
\usage{
\method{confint}{mbm}(x, parm = "fits", level = 0.95,
  method = c("parametric", "sample"))
}
\arguments{
\item{parm}{Specification of what parameters to consider. Currently, only `fits` is 
supported; this produces confidence intervals for the predictions of the original
data.}

\item{level}{Confidence level required}

\item{method}{How to compute confidence interval. The default uses a parametric estimate;
this is the only method currently implemented.}

\item{object}{An MBM object}
}
\value{
A matrix with upper and lower confidence limits
}
\description{
Confidence intervals for MBM models
}
