% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwm.r
\name{cwm}
\alias{cwm}
\title{Community weighted mean trait values}
\usage{
cwm(comm, traits, na.rm = FALSE)
}
\arguments{
\item{comm}{Site by species weights matrix; if presence-absence (1-0), the function returns the community mean}

\item{traits}{Matrix of traits; rownames will be matched to colnames in \code{comm}}

\item{na.rm}{Logical, whether to remove NAs when computing the mean}
}
\value{
site by trait CWM matrix
}
\description{
Community weighted mean trait values
}
