% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dendro.r
\name{dendrogram}
\alias{dendrogram}
\title{Build a functional dendrogram to be used with tree-based diversity indices}
\usage{
dendrogram(x, vartypes, species = 1, traits = 2, values = 3)
}
\arguments{
\item{x}{"Tall" dataframe with species and trait information; see 'details'}

\item{vartypes}{2-column data frame or character matrix giving (in the first column) the trait names and (in the
second column) the trait types. Values in the second column should be one of 'Q' (quantitative), 'N' (nominal
or categorical), or 'O' (ordinal).}

\item{species}{column number or name in x containing species names, or a vector of species names}

\item{traits}{column number or name in x of trait names, or a vector of trait names}

\item{values}{column or name of trait values, or vector of trait values; see 'details'}
}
\value{
named list containing a species by species trait distance matrix and a functional dendrogram (in phylo format)
}
\description{
Build a functional dendrogram to be used with tree-based diversity indices
}
\details{
If 'x' is missing, then species, traits, and values must be vectors containing the data
}

