% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{dendrosor}
\alias{dendrosor}
\title{Sorensen dissimilarity index based on a dendrogram}
\usage{
dendrosor(comm, distmat, partition = TRUE)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns).}

\item{distmat}{Distance matrix; margins must match exactly the columns in comm}
}
\value{
If \code{partition == TRUE}, list of three matrices giving the turnover,
		nestedness, and total components of dissimilarity. Otherwise, a matrix of
		pairwise dissimilarity values between communities.
}
\description{
Produces dissimilarity matrices using a species by species distance matrix; useful
for example for functional diversity.
}
\references{
Pavoine, S. and Ricotta, C. 2014. Functional and phylogenetic similarity 
		among communities. Methods in Ecology and Evolution 5(7): 666–675.
}
