% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm_io.r
\name{fix_slashes}
\alias{fix_slashes}
\title{Standardize slashes to prevent file errors on windows systems}
\usage{
fix_slashes(fname)
}
\arguments{
\item{fname}{A vector of file names}
}
\value{
Modified filenames
}
\description{
Standardize slashes to prevent file errors on windows systems
}
\keyword{internal}
