% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{gp_params}
\alias{gp_params}
\title{GP parameters for an MBM object}
\usage{
gp_params(x)
}
\arguments{
\item{An}{\code{\link{mbm}} object}
}
\value{
A list with 2 named items; \code{$mean} is the mean vector and \code{$cholesky}
		is the Cholesky decomposition of the variance-covariance matrix of the SVGP.
}
\description{
GP parameters for an MBM object
}
