% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{inducing}
\alias{inducing}
\title{Matrix of inducing inputs for an MBM object}
\usage{
inducing(x)
}
\arguments{
\item{An}{\code{\link{mbm}} object}
}
\value{
A matrix of inducing inputs
}
\description{
Matrix of inducing inputs for an MBM object
}
