% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.r
\name{jaccard}
\alias{jaccard}
\title{Compute jaccard dissimilarity}
\usage{
jaccard(x, community, site_col = 1, species_col = 2, pres_abs = NULL)
}
\arguments{
\item{x}{a data frame. Minimally containing a column for site names, a column for species names present at
each site, and a column for abundances}

\item{community}{optional site by species presence-absence matrix. Site names, if desired, can be specified as the row names. 
If community is specified, then x is optional}

\item{sites}{column name or index of x containing site names}

\item{species}{column name or index of x containing species names}

\item{abundance}{column name or index of abundance column}
}
\value{
Site by site dissimilarity matrix
}
\description{
Compute jaccard dissimilarity
}
\details{
If both x and community are provided, x will be used
}
