% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm.r
\name{make_args}
\alias{make_args}
\title{Produce arguments for the python call}
\usage{
make_args(x, files, GPy_location = NA, n_samples = NA,
  tfOutput = "_out.csv")
}
\arguments{
\item{x}{An MBM object}

\item{files}{Character vector of filenames generated from \link{\code{write_mbm_dat}}}
}
\value{
A character vector of arguments to a python call
}
\description{
Produce arguments for the python call
}
\keyword{internal}
