% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{make_mbm}
\alias{make_mbm}
\title{Construct (but do not fit) an mbm model}
\usage{
make_mbm(y, x, y_name, link, likelihood, lengthscale, sparse,
  force_increasing, sparse_inducing = 10, sparse_batch = 10,
  sparse_iter = 10000)
}
\arguments{
\item{y}{Square dissimilarity or distance matrix, can be complete or lower triangular 
only. Row and column names are required and must match the site names in the 
rows of \code{x}.}

\item{x}{Matrix giving a series of covariates (in columns) for all sites (in rows). 
Row names are required. All variables will be included in the model.}

\item{y_name}{A name to give to the y variable}

\item{link}{Link function to use}

\item{likelihood}{Likelihood function to use}

\item{lengthscale}{Either missing (in which case all lengthscales will be optimized) or
a numeric vector of length \code{ncol(x)+1}. If a vector, the first entry 
corresponds to environmental distance, and entries \code{i = 1 + (1:n)} 
correspond to the variable in x[,i]. Values must be \code{NA} or positive 
numbers; if NA, the corresponding lengthscale will be set via optimization, 
otherwise it will be fixed to the value given.}

\item{sparse}{Should we use the stochastic variational GP.}

\item{force_increasing}{Boolean; if true, beta diversity will be constrained to 
increase with environmental distance}

\item{sparse_inducing}{Number of inducing inputs to use for the svgp}

\item{sparse_batch}{Batchsize to use for the svgp}

\item{sparse_iter}{Maximum number of optimizer iterations for svgp}
}
\value{
An mbm object
}
\description{
Construct (but do not fit) an mbm model
}
\keyword{internal}
