% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbm.r
\name{make_symmetric}
\alias{make_symmetric}
\title{Turn an MBM prediction dataframe into a symmetric matrix}
\usage{
make_symmetric(DF, formula, value.var, ...)
}
\arguments{
\item{DF}{MBM predfiction dataframe}

\item{formula}{A formula to be passed to \code{link{reshape2::acast}}}

\item{value.var}{Name of value variable for \code{acast}}

\item{...}{Additional parameters for \code{acast}}
}
\value{
A symmetric matrix of predictions
}
\description{
Turn an MBM prediction dataframe into a symmetric matrix
}
\keyword{internal}
