% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.r
\name{match_mat_dims}
\alias{match_mat_dims}
\title{Match matrix dimensions}
\usage{
match_mat_dims(x, y, by.x = c("r", "c", "rc"), by.y = c("r", "c"))
}
\arguments{
\item{x, y}{matrices to match}

\item{by.x}{should we match rows, columns, or both?}
}
