% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmfit.r
\name{mbmfit}
\alias{mbmfit}
\title{MBM Fit Class}
\usage{
mbmfit(fit_file, result_dir, holdout_file, predict_file,
  link_function = c("identity", "log", "probit"),
  model_name = basename(result_dir), type = c("alpha", "beta"))
}
\arguments{
\item{fit_file}{File name for the data used to fit the model}

\item{result_dir}{Directory with results}

\item{holdout_file}{(optional) File name for the holdout dataset}

\item{predict_file}{(optional) File name for the prediction dataset}

\item{link_function}{Link function used to fit the model}

\item{model_name}{What to call the model}

\item{type}{Alpha or beta diversity?}
}
\value{
An object of class mbmfit
}
\description{
Creates and object from an mbm fit (from Python)
}

