% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{plot.mbm}
\alias{plot.mbm}
\alias{print.mbm}
\alias{print.mbm}
\alias{summary.mbm}
\alias{is.mbm}
\alias{format.mbm}
\title{Standard R methods for mbm objects}
\usage{
plot.mbm(x, line = TRUE, sterr = FALSE, ...)

\method{print}{mbm}(x)

\method{summary}{mbm}(x)

is.mbm(x)

\method{format}{mbm}(x)
}
\arguments{
\item{x}{An \code{\link{mbm}} object}

\item{line}{Boolean; should we show the 1:1 line?}

\item{sterr}{Boolean; if true, standard errors will be shown with the points}

\item{...}{Additional arguments to be passed to base graphics plotting commands}
}
\description{
Standard R methods for mbm objects
}
