% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.r
\name{mpd}
\alias{mpd}
\title{Mean pairwise distance}
\usage{
mpd(community, phylogeny, traits, type = c("a", "b", "ab"), ...)
}
\arguments{
\item{community}{Community matrix (required), giving sites (rows) by species (columns). Column names should match the
names in the phylogeny, if supplied.}

\item{phylogeny}{Phylogeny (optional) in the format used by \link{\code{ape}}; if supplied, phylogenetic diversity
will be computed.}

\item{traits}{Matrix of species (rows) by trait values (columns); traits may be a mix of quantitative
and categorical.}

\item{type}{What kind of diversity to compute; either alpha (a), beta (b) or both (ab)}

\item{...}{Additional arguments to pass to \code{picante} functions.}
}
\value{
If a single index is selected a vector(for alpha) or matrix (for beta) with the selected diversity metric.
         If multiple indices are selected, a list of diversity vectors/matrices
}
\description{
Mean pairwise distance
}
\details{
This is a generic wrapper around several functions from the \link{\code{picante}} package. It has facilities
          for computing mpd for both alpha and beta diversity, for functional and phylogenetic diversity.
}

