% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm.r
\name{parse_param_names}
\alias{parse_param_names}
\title{Determine parameter names to a fit mbm model}
\usage{
parse_param_names(x)
}
\arguments{
\item{x}{A fit mbm model}
}
\value{
A vector of parameter names that can be assigned to names(x$params)
}
\description{
Determine parameter names to a fit mbm model
}
\keyword{internal}
