% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{phylosor}
\alias{phylosor}
\title{Phylogenetic sorensen dissimilarity index}
\usage{
phylosor(comm, tree, partition = FALSE)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns).}

\item{tree}{Dendrogram of class \code{phylo}}

\item{partition}{Logical; if true, the index will be partitioned into turnover and
nestedness.}
}
\value{
If \code{partition == TRUE}, list of three matrices giving the turnover,
		nestedness, and total components of dissimilarity. Otherwise, a matrix of
		pairwise dissimilarity values between communities.
}
\description{
Phylogenetic sorensen dissimilarity index
}
\references{
Pavoine, S. and Ricotta, C. 2014. Functional and phylogenetic similarity 
		among communities. Methods in Ecology and Evolution 5(7): 666–675.
}
