% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mbm.r
\name{plot.mbm}
\alias{plot.mbm}
\title{Plot predictions vs fits for an mbm object}
\usage{
\method{plot}{mbm}(x, line = TRUE, sterr = FALSE, ...)
}
\arguments{
\item{x}{An \code{\link{mbm}} object}

\item{line}{Boolean; should we show the 1:1 line?}

\item{sterr}{Boolean; if true, standard errors will be shown with the points}

\item{...}{Additional arguments to be passed to base graphics plotting commands}
}
\description{
Plot predictions vs fits for an mbm object
}
