% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbm.r
\name{predict.mbm}
\alias{predict.mbm}
\title{Predict method for MBM objects}
\usage{
\method{predict}{mbm}(x, newdata)
}
\arguments{
\item{x}{A previously-fit MBM object}

\item{newdata}{Optional dataset for prediction. If present, it should be a new dataset in 
the same format used to fit the model (i.e., a site by covariate matrix). If missing,
predictions will be for the original data.}

\item{n_samples}{NA or integer; if NA, analytical predictions with standard deviation 
are returned, otherwise posterior samples are returned.}
}
\value{
A data frame of predictions and standard deviations (on the link scale); use 
		\code{x$y_rev_transform(x$rev_link(predictions$fit))} for the response scale.
}
\description{
Predict method for MBM objects
}
\details{
Prediction to new data is possible after the fact for mbm models, however 
		there can be performance penalties for doing so with large models. Thus, it is 
	sometimes preferable to predict during model fitting via the \code{predictX} 
		argument to the \code{\link{mbm}} function. 

		All prediction is done on the link scale.

	This function caches to disk, thus it is important to ensure that adequate disk 
		space is available when using large prediction datasets.
}
