% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbm.r
\name{predict_mbm_raster}
\alias{predict_mbm_raster}
\title{Prediction for MBM from a raster dataset}
\usage{
predict_mbm_raster(x, rasterdat, ...)
}
\arguments{
\item{x}{A previously-fit MBM object}

\item{rasterdat}{Raster stack containing named layers matching the variable names in x (i.e., colnames(x$covariates)[-1]).
If a layer named 'names' is included, this layer will be used as sitenames, otherwise they will be assigned unique
numbers}

\item{...}{Other named parameters to pass to \code{\link{predict.mbm}}.}
}
\value{
A named list; \code{fits} is a cell by cell matrix of predictions (on the link scale; use \code{x$y_rev_transform(x$rev_link(predictions$fit))} 
   for the response scale), \code{stdev} is a cell by cell matrix of standard deviations, and \code{coords} is a matrix of coordinates. Row/column
   names in \code{fits} and \code{stdev} match the rownames in \code{coords}.
}
\description{
Prediction for MBM from a raster dataset
}
