% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm.r
\name{prep_ls}
\alias{prep_ls}
\title{Convenience function to set up the lengthscale for passing to python}
\usage{
prep_ls(lengthscale)
}
\arguments{
\item{lengthscale}{A \code{lengthscale} from an \code{\link{mbm}} object}
}
\value{
A lengthscale string suitable for passing to python
}
\description{
Convenience function to set up the lengthscale for passing to python
}
\keyword{internal}
