% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{prep_predict}
\alias{prep_predict}
\title{Set up a prediction dataset to pass to python}
\usage{
prep_predict(newdata, x, ...)
}
\arguments{
\item{newdata}{a dataset for prediction}

\item{x}{an MBM object}

\item{...}{Additional arguments to be passed to \link{\code{env_dissim}}}
}
\value{
A processed dataset to send to MBM
}
\description{
Set up a prediction dataset to pass to python
}
\keyword{internal}
