% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.r
\name{rc}
\alias{rc}
\title{Plot response curves}
\usage{
rc(x, missing_action = c("error", "empirical", "fit"),
  rc_name = "rc_distance", col_pt = "#666666", col_line = "#ce5336",
  col_polygon = paste0(col_line, "66"), cex_pt = 0.4, add = FALSE,
  ...)
}
\arguments{
\item{x}{An MBM object}

\item{missing_action}{What to do if no response curve was fit with the model. \code{empirical} will make an empirical
curve from the model fits. \code{fit} will fit a response curve, which can be slow. \code{error} (the only implemented
option) quits with an error message.}

\item{rc_name}{A character string, integer, or NA. If a string or integer, the name or index of the response curve to fit. If NA,
(not yet implemented), will attempt to find all response curves and plot them all.}

\item{col_pt}{Point color, if NA points will not be plotted.}

\item{col_line}{Line color, if NA line will not be plotted. Note that a hex color name must be used if col_polygon is not specified}

\item{col_polygon}{Polygon color; if NA polygon will not be plotted}

\item{cex_pt}{Size of point plotting symbols}

\item{add}{boolean; if true this plot will be added to the current plot}

\item{...}{Additional arguments to be passed to base graphics plotting commands}
}
\description{
Produce a plot of a response curve from an MBM object, with some sensible defaults.
If a response curve was not generated when the model was fit, teh default is to try to make an empirical curve
from the fits.
}
