% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm_io.r
\name{read_mbm_predict}
\alias{read_mbm_predict}
\title{Higher-level function for reading predictions}
\usage{
read_mbm_predict(fname, nsamp = NA, tfExt = ".csv",
  tfOutput = ".out", nameExt = ".names")
}
\arguments{
\item{fname}{The name of the input file used to generate the predictions}

\item{nsamp}{Number of samples, if used}

\item{tfExt}{Extension used for input files}

\item{tfOutput}{file extension for output files}

\item{namesExt}{What should be appended to the names filename}
}
\value{
A data frame of predictions and site names
}
\description{
Reads all predictions and site names if available
}
\keyword{internal}
