% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.r
\name{richness}
\alias{richness}
\title{Compute species richness}
\usage{
richness(x)
}
\arguments{
\item{x}{A matrix with rows as sites and species as columns. Values are abundances or presence/absence}
}
\value{
A vector of diversity values
}
\description{
Compute species richness
}
