% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{set_link}
\alias{set_link}
\title{Add \code{link} and \code{rev_link} closures to mbm model}
\usage{
set_link(x, link = c("identity", "probit", "log"))
}
\arguments{
\item{x}{an mbm object}

\item{link}{character; which link function to use}
}
\value{
\code{x} x with link and rev_link functions added
}
\description{
If a mean function is used, the link will be set to identity, and y_transform and y_untransform
methods will also be added.
}
\keyword{internal}
