% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{set_mbm_link}
\alias{set_mbm_link}
\title{Set up link functions for mbm objects}
\usage{
set_mbm_link(x, link)
}
\arguments{
\item{x}{mbm model object}

\item{link}{Link function to use}
}
\value{
A copy of the mbm object with the link function set
}
\description{
This is the only supported way for changing link functions in mbm objects; do not try
to do it by hand
}
\keyword{internal}
