% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{set_mean_function}
\alias{set_mean_function}
\title{Set up MBM mean function}
\usage{
set_mean_function(dim, useMeanFunction)
}
\arguments{
\item{dim}{Kernel dimension (number of variables)}

\item{useMeanFunction}{Logical, should the mean function be used?}
}
\description{
Set up MBM mean function
}
\keyword{internal}
