% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{set_ytrans}
\alias{set_ytrans}
\title{Set y transformations for an MBM object assuming a linear increasing mean function}
\usage{
set_ytrans(x, ydat, link, eps = 0.001)
}
\arguments{
\item{x}{an MBM object}

\item{ydat}{vector of y data points}

\item{link}{character, link function to use}

\item{eps}{Constant to add to avoid infinite values for probit}
}
\value{
An mbm object
}
\description{
Set y transformations for an MBM object assuming a linear increasing mean function
}
\keyword{internal}
