% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{setup_mbm_kernel}
\alias{setup_mbm_kernel}
\title{Set up MBM kernel}
\usage{
setup_mbm_kernel(dim, lengthscale = NULL, prior, sparse,
  which = c("all", "lengthscale", "variance"))
}
\arguments{
\item{dim}{Kernel dimension (number of variables)}

\item{lengthscale}{Lengthscale parameter as from [mbm()]. 
    Either NULL (in which case all lengthscales will be optimized) or 
a numeric vector of length \code{ncol(x)+1}. If a vector, the first entry 
    corresponds to environmental distance, and entries \code{i = 1 + (1:n)} 
    correspond to the variable in x[,i]. Values must be \code{NULL} or positive 
    numbers; if NULL, the corresponding lengthscale will be set via optimization, 
    otherwise it will be fixed to the value given.}

\item{prior}{Prior distribution to use; currently ignored}

\item{sparse}{Logical, should a sparse GP be used?}

\item{which}{Which parameters to set up, either all, variance params, or lengthscale}
}
\description{
Set up MBM kernel
}
\keyword{internal}
