% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mbm.r
\name{setup_y_transform}
\alias{setup_y_transform}
\title{Set y transformations for an MBM object}
\usage{
setup_y_transform(x, link, eps = 0.001)
}
\arguments{
\item{x}{an MBM object}

\item{link}{character, link function to use}

\item{eps}{Constant to add to avoid infinite values for probit}
}
\description{
Set y transformations for an MBM object
}
\details{
Sets up y transformations to use instead of a link function for probit links. When
		the y-data include zeros and ones, an additional step is necessary to squeeze the data
		from [0,1], (0,1], or [0,1) to the open interval (0,1). In the [0,1] case, the smithson
		lemon-squeezer is used: y􏰂' = [y􏰀(N – 1) + 1/2]/N. Otherwise, eps is first added or
		subtracted from all values.
}
\references{
Smithson, M. and Verkuilen, J. 2006. A Better Lemon Squeezer? Maximum-Likelihood
	Regression With Beta-Distributed Dependent Variables. Psychological Methods 11(1): 54-71.
}
\keyword{internal}
