% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.r
\name{shannon}
\alias{shannon}
\title{Shannon diversity}
\usage{
shannon(x)
}
\arguments{
\item{x}{A matrix with rows as sites and species as columns. Values are proportional abundance}
}
\value{
A vector of diversity values
}
\description{
Shannon diversity
}
