% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.r
\name{simpson}
\alias{simpson}
\title{Compute simpson diversity}
\usage{
simpson(x, proportion = TRUE)
}
\arguments{
\item{x}{A matrix with rows as sites and species as columns. Values are abundances}

\item{proportion}{Logical. If true, abundances in x are considered proportional abundances. If false, x will be
transformed to proportional abundance}
}
\value{
A vector of diversity values
}
\description{
Compute simpson diversity
}

