% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.r
\name{smithson}
\alias{rev_smithson}
\alias{smithson}
\title{Smithson transformation}
\usage{
smithson(x)

rev_smithson(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Transformation of x
}
\references{
Smithson M and Verkuilen J. 2006. A Better Lemon Squeezer? Maximum-Likelihood
              Regression With Beta-Distributed Dependent Variables. Psych Meth 11:54-71.
}

