% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{sor}
\alias{sor}
\title{Sorensen dissimilarity (without distance)}
\usage{
sor(comm)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns).}
}
\value{
Matrix of pairwise dissimilarity values between communities.
}
\description{
Compute original sorensen dissimilarity index, with no option for a distance matrix among species
}
\details{
The index is defined as 1 - (2 * (A ∩ B) / (A + B)).
}
\references{
Sørensen, T. (1948) A method of establishing groups of equal amplitude in plant sociology 
based on similarity of species content. Kongelige Danske Videnskabernes Selskabs Biologiske Skrifter, 5, 1–34.
}
