% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{sorensen}
\alias{sorensen}
\title{Sorensen dissimilarity index}
\usage{
sorensen(comm, dis)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns).}

\item{dist}{Optional species by species distance matrix; margin names must match colnames in comm.}
}
\value{
Matrix of pairwise dissimilarity values between communities.
}
\description{
Sorensen dissimilarity index
}
\references{
Pavoine, S. and Ricotta, C. 2014. Functional and phylogenetic similarity among communities.
Methods in Ecology and Evolution 5(7): 666–675.
}
