% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen_pd.r
\name{sorensen_pd}
\alias{sorensen_pd}
\title{Sorensen phylogenetic distance}
\usage{
sorensen_pd(communities, phylogeny, na.species.action = c("error", "omit"))
}
\arguments{
\item{phylogeny}{Phylogeny in the format used by \link{\code{ape}}}

\item{na.species.action}{Action to take if species are found in the community matrix that are missing in the
phylogeny. See details.}

\item{community}{Community matrix giving sites (rows) by species (columns). Column names should match the
names in the phylogeny}
}
\value{
Matrix of pairwise dissimilarity values between communities
}
\description{
Sorensen phylogenetic distance
}
\details{
Compute an analogue of the sorensen distance, using shared pd instead of taxonomic diversity

If na.species.action is set to 'omit,' missing species will be dropped from the community matrix
}

