% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{plot.mbmSP}
\alias{plot.mbmSP}
\alias{print.mbmSP}
\alias{summary.mbmSP}
\alias{is.mbmSP}
\title{Standard R methods for mbmSP objects}
\usage{
\method{plot}{mbmSP}(x, sterr = FALSE, ...)

\method{print}{mbmSP}(x)

\method{summary}{mbmSP}(x)

is.mbmSP(x)
}
\arguments{
\item{x}{An \code{\link{mbmSP}} object}

\item{sterr}{Boolean; if true, standard errors will be mapped instead of fits}

\item{...}{Additional arguments to be passed to plotting commands}
}
\description{
Standard R methods for mbmSP objects
}
