% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm_io.r
\name{write_mbm_dat}
\alias{write_mbm_dat}
\title{Write MBM data to disk}
\usage{
write_mbm_dat(x, dir = tempdir(), tfBase = "mbm_", tfExt = ".csv",
  namesExt = ".names")
}
\arguments{
\item{x}{An MBM object}

\item{dir}{Directory to which to write the files}

\item{tfBase}{Base name for the temp files}

\item{tfExt}{Extension for temp files}

\item{namesExt}{What should be appended to the names filename}
}
\value{
A named character vector with the file names for the \code{response},
    \code{covariates}, and \code{params}. As a side effect, the response,
    covariate, and (if present) params are written to the files.
}
\description{
Write MBM data to disk
}
\keyword{internal}
