% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm_io.r
\name{write_mbm_predict}
\alias{write_mbm_predict}
\title{Write prediction data to disk}
\usage{
write_mbm_predict(x, datname, tfBase = "mbm_", tfExt = ".csv",
  bigLim = 200, namesExt = ".names")
}
\arguments{
\item{x}{Prediction dataset, including site names in the last two columns}

\item{datname}{The name of the prediction set}

\item{bigLim}{Integer, what's the largest number of sites before we start treating this as a big data problem}

\item{namesExt}{What should be appended to the names filename}
}
\value{
a character vector; the first element is a string for the file name, the second is an argument for passing
     to the mbm python program; as a side effect, the file is written; a companion file is written with just site names
}
\description{
Write prediction data to disk
}
\keyword{internal}
