% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.r
\name{alpha_diversity}
\alias{alpha_diversity}
\alias{richness}
\alias{simpson}
\alias{shannon}
\title{Compute alpha diversity metrics}
\usage{
richness(x)

simpson(x, proportion = TRUE)

shannon(x)
}
\arguments{
\item{x}{A matrix with rows as sites and species as columns. Values are abundances or presence/absence}

\item{proportion}{Logical. If true, abundances in x are considered proportional abundances. If false, x will be
transformed to proportional abundance}
}
\value{
A vector of diversity values
}
\description{
Compute alpha diversity metrics
}
