% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissim.r
\name{bc}
\alias{bc}
\alias{jaccard}
\title{Compute dissimilarity indices}
\usage{
bc(comm)

jaccard(comm)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns); values are abundances. Column and row labels are highly recommended}
}
\value{
Site by site dissimilarity matrix
}
\description{
Compute dissimilarity indices
}
