% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_mbm.r
\name{confint.mbm}
\alias{confint.mbm}
\title{Confidence intervals for MBM models}
\usage{
\method{confint}{mbm}(object, parm, level = 0.95, method = c("auto",
  "parametric", "sample"))
}
\arguments{
\item{object}{An MBM object}

\item{parm}{Specification of what parameters to consider. Default \code{'fits'} or a value of 0 gives confidence intervals for fitted data points.
Other strings (or numerical index) will be looked up in the model predictions.}

\item{level}{Confidence level required}

\item{method}{How to compute confidence interval. \code{'auto'} (the default) uses parametric estimates of the confidence intervals;
this is the only method currently implemented.}

\item{col_line}{Line color}

\item{col_polygon}{Polygon color}

\item{cex_pt}{Size of point plotting symbols}

\item{...}{Additional arguments to be passed to base graphics plotting commands}
}
\value{
A matrix with upper and lower confidence limits
}
\description{
Produce a plot of a response curve from an MBM object, with some sensible defaults.
If a response curve was not generated when the model was fit, teh default is to try to make an empirical curve
from the fits.
}
