% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_dissim.r
\name{env_dissim}
\alias{env_dissim}
\title{Compute a dataframe of environmental dissimilarity values}
\usage{
env_dissim(x, sites = 0, scale = TRUE, sort = TRUE)
}
\arguments{
\item{x}{A data frame of environmental covariates.}

\item{sites}{Either a vector of site names or a column index (or column name) giving the location of the site names
in x. If 0, row names will be used}

\item{scale}{A boolean; if true, computations will be done on the scaled site}

\item{sort}{Logical, should we sort by site names?}
}
\value{
A data frame including both sites' names, environmental dissimilarity between the two sites, and midpoints
        between the two sites for each variable
}
\description{
Compute a dataframe of environmental dissimilarity values
}

