% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm.r
\name{mbm}
\alias{mbm}
\title{Create an mbm model}
\usage{
mbm(y, x, predictX, scale = TRUE, y_name = "beta")
}
\arguments{
\item{y}{Square dissimilarity or distance matrix, can be complete or lower triangular only. Row and column names are
required and must match the site names in the rows of \code{x}.}

\item{x}{Matrix giving a series of covariates (in columns) for all sites (in rows). Row names are required. All 
variables will be included in the model.}

\item{predictX}{List of prediction datasets; each list element is a matrix with same number of columns as \code{x}. See details.}

\item{scale}{Boolean, if true, x values will be centered and scaled before fitting the model.}

\item{y_name}{A name to give to the y variable}
}
\value{
An S3 object of class mbm.
}
\description{
Create an mbm model
}
\details{
Prediction datasets can either be supplied when the model is called, or by using the \code{predict} method on the \code{mbm}
         object. The former will generally be much faster to run; see \code{\link{predict.mbm}}. Note that predictions are always
         generated for the input dataset.
}
