% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm.r
\name{mbm}
\alias{mbm}
\title{Create an mbm model}
\usage{
mbm(y, x, y_name = "beta", pyMsg = FALSE)
}
\arguments{
\item{y}{Square dissimilarity or distance matrix, can be complete or lower triangular 
only. Row and column names are required and must match the site names in the 
rows of \code{x}.}

\item{x}{Matrix giving a series of covariates (in columns) for all sites (in rows). Row 
names are required. All variables will be included in the model.}

\item{y_name}{A name to give to the y variable}

\item{pyMsg}{boolean, should we print messages from python? Useful for debugging}

\item{link}{Link function to use}

\item{predictX}{List of prediction datasets; each list element is a matrix with same 
number of columns as \code{x}. See details.}

\item{scale}{Boolean, if true, x values will be centered and scaled before fitting the
model.}

\item{n_samples}{NA or integer; if NA, analytical predictions with standard deviation 
are returned, otherwise posterior samples are returned}

\item{response_curve}{The type of response curve to generate. The default 
(\code{distance}) will predict over a range of distances assuming pairs of 
sites equally spaced across the midpoint of environmental space. \code{none} 
Produces no response curve, while \code{all} creates response curves for all
variables.}

\item{lengthscale}{Either missing (in which case all lengthscales will be optimized) or 
a numeric vector of length \code{ncol(x)+1}. If a vector, the first entry 
    corresponds to environmental distance, and entries \code{i = 1 + (1:n)} 
    correspond to the variable in x[,i]. Values must be \code{NA} or positive 
    numbers; if NA, the corresponding lengthscale will be set via optimization, 
    otherwise it will be fixed to the value given.}

\item{force_increasing}{Boolean; if true, beta diversity will be constrained to 
increase with environmental distance}

\item{svgp}{Should we use the stochastic variational GP (see 'details').}

\item{exact_thresh}{integer; threshold at which mbm will refuse to run an exact gp.}

\item{...}{Additional arguments to pass to MBM}
}
\value{
An S3 object of class mbm.
}
\description{
Create an mbm model
}
\details{
Prediction datasets can either be supplied when the model is called, or by 
			using the \code{predict} method on the \code{mbm} object. The former will 
			generally be much faster to run; see \code{\link{predict.mbm}}. Note that 
			predictions are always generated for the input dataset.

			For larger datasets (more than ~100 sites), it is recommended to use
			\code{svgp=TRUE}. This will use a sparse approximation to the default
			method, following the stochastical variational GP (Hensman et al 2013).
			Note that if a link function is selected, it will be applied as a 
			transformation of the y data--i.e., for link function L() we fit a SVGP to
			describe the expectation E(L(y))--rather than as a true link function--
			fitting L(E(y))--as is done when \code{svgp=FALSE}. This is due to a 
			limitation in the underlying GP library.
}
\references{
Hensman J, Fusi N, and Lawrence ND. 2013. Gaussian Processes for Big Data.
			In: In Proceedings of the 29th Conference on Uncertainty in Artificial 
			Intelligence.
}
