% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbm.r
\name{mbm}
\alias{mbm}
\title{Create an mbm model}
\usage{
mbm(y, x, predictX, link = c("identity", "probit", "log"), scale = TRUE,
  n_samples = NA, response_curve = c("distance", "none", "all"),
  lengthscale, y_name = "beta", GPy_location, pyCmd = "python",
  pyMsg = FALSE)
}
\arguments{
\item{y}{Square dissimilarity or distance matrix, can be complete or lower triangular only. Row and column names are
required and must match the site names in the rows of \code{x}.}

\item{x}{Matrix giving a series of covariates (in columns) for all sites (in rows). Row names are required. All 
variables will be included in the model.}

\item{predictX}{List of prediction datasets; each list element is a matrix with same number of columns as \code{x}. See details.}

\item{link}{Link function to use}

\item{scale}{Boolean, if true, x values will be centered and scaled before fitting the model.}

\item{n_samples}{NA or integer; if NA, analytical predictions with standard deviation are returned, otherwise posterior samples are returned}

\item{response_curve}{The type of response curve to generate. The default (\code{distance}) will predict over a range of distances
assuming pairs of sites equally spaced across the midpoint of environmental space. \code{none} Produces no response curve,
while \code{all} creates response curves for all variables.}

\item{lengthscale}{Either missing (in which case all lengthscales will be optimized) or a numeric vector of length \code{ncol(x)+1}.
If a vector, the first entry corresponds to environmental distance, and entries \code{i = 1 + (1:n)} correspond to the variable in 
x[,i]. Values must be \code{NA} or positive numbers; if NA, the corresponding lengthscale will be set via optimization, otherwise
it will be fixed to the value given.}

\item{y_name}{A name to give to the y variable}

\item{GPy_location}{Optional character giving the location of the user's GPy installaion}

\item{pyCmd}{Where to look for python; the version in use must have GPy installed}

\item{pyMsg}{boolean, should we print messages from python? Useful for debugging}
}
\value{
An S3 object of class mbm.
}
\description{
Create an mbm model
}
\details{
Prediction datasets can either be supplied when the model is called, or by using the \code{predict} method on the \code{mbm}
         object. The former will generally be much faster to run; see \code{\link{predict.mbm}}. Note that predictions are always
         generated for the input dataset.
}
