% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{plot.mbm}
\alias{plot.mbm}
\alias{print.mbm}
\alias{coef.mbm}
\alias{print.mbm}
\alias{summary.mbm}
\alias{is.mbm}
\alias{format.mbm}
\title{Standard R methods for mbm objects}
\usage{
\method{plot}{mbm}(x, line = TRUE, sterr = FALSE, ...)

\method{coef}{mbm}(x)

\method{print}{mbm}(x)

\method{summary}{mbm}(x)

is.mbm(x)

\method{format}{mbm}(x)
}
\arguments{
\item{x}{An \code{\link{mbm}} object}

\item{line}{Boolean; should we show the 1:1 line?}

\item{sterr}{Boolean; if true, standard errors will be shown with the points}

\item{...}{Additional arguments to be passed to base graphics plotting commands}
}
\description{
Standard R methods for mbm objects
}
\seealso{
For mbm objects fit with \code{svgp=TRUE}, \code{\link{inducing}} for the 
		matrix of inducing inputs, \code{\link{gp_params}} for the mean vector
		and the variance-covariance matrix of the GP
}
