% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{sor}
\alias{sor}
\title{Sorensen dissimilarity (without partitioning)}
\usage{
sor(comm)
}
\arguments{
\item{comm}{Community matrix giving sites (rows) by species (columns).}
}
\value{
Matrix of pairwise dissimilarity values between communities.
}
\description{
Compute sorensen dissimilarity index; deprecated method (but faster than new one)
}
\details{
The index is defined as 1 - (2 * (A ∩ B) / (A + B)).
}
\keyword{internal}
