% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen.r
\name{sorensen}
\alias{sorensen}
\title{Sorensen phylogenetic distance}
\usage{
sorensen(community, binomial = FALSE)
}
\arguments{
\item{community}{Community matrix giving sites (rows) by species (columns). Column names should match the
names in the phylogeny}

\item{binomial}{Should we return the data in binomial format (matrix of X successes and matrix of N trials)}
}
\value{
If binomial is FALSE, matrix of pairwise dissimilarity values between communities. If binomial is TRUE, 
            list of matrices, the first being the numerator of the sorensen index and the second being the denominator
}
\description{
Sorensen phylogenetic distance
}
\details{
The index is defined as 1 - (2 * (A ∩ B) / (A + B)). For binomial data, we return (A + B) - (2 * (A ∩ B)) as the 
success matrix, so that a binomial regression will make predictions on the dissimilarity (rather than the similarity)
}
