% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mbm.r
\name{spatial_predict}
\alias{spatial_predict}
\title{Spatial MBM prediction}
\usage{
spatial_predict(x, rasterdat, stdev = FALSE, ...)
}
\arguments{
\item{x}{A previously-fit MBM object}

\item{rasterdat}{Raster stack containing named layers matching the variable names in x (i.e., colnames(x$covariates)[-1])}

\item{stdev}{Boolean; should the standard deviation be included}

\item{...}{Other named parameters to pass to \code{\link{predict.mbm}}.}
}
\value{
A raster stack with 3 or 4 (if stdev==TRUE) layers; the first 3 layers can be interpreted as RGB channels for
    visualization with plotRGB, the fourth is the standard deviation.
}
\description{
Spatial MBM prediction
}
\details{
Notes about this function forthcoming. If \code{stdev == TRUE}, the mean standard deviation for each raster cell
    will be computed
}
